Mass tidy up of minor compilation warning issues.
These cover unused variables, variable names shadowing "parent" ones,
missing prototypes, enum lists ending with comma, etc.


git-svn-id: https://staden.svn.sourceforge.net/svnroot/staden/staden/trunk@2126 ba6f2710-ca9f-427c-89d3-5bf716dbc8e1

+++ src/spin/seq_reg.c
@@ -641,7 +641,7 @@ void seq_register_dump(void) {
     for (c = 0; c < ArrayMax(sequence_reg); c++) {
 
 	printf("sequence %d\n", c);
-	printf("num funcs!! %d \n", seq_Nfuncs(c));
+	printf("num funcs!! %d \n", (int)seq_Nfuncs(c));
 	n = seq_Nfuncs(c);
 	r = ArrayBase(seq_reg, seq_func_array(c));
